api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile api/requirements.in
#
certifi==2019.6.16
chardet==3.0.4
click==7.0                # via pip-tools
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
django-datetime-utc==1.0.2
django-debug-toolbar==2.0
django-extensions==2.1.9
django-filter==2.2.0
django==2.2.3
djangorestframework-csv==2.1.0
djangorestframework-gis==0.14
djangorestframework-xml==1.4.0
djangorestframework==3.10.0
drf-amsterdam==0.3.2
drf-extensions==0.5.0
drf-flex-fields==0.5.0
drf-yasg==1.16.1
factory-boy==2.12.0
faker==2.0.0
graypy==1.1.3
idna==2.8
inflection==0.3.1
itypes==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
packaging==19.0
pip-review==1.0
pip-tools==3.8.0
psycopg2-binary==2.8.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.100
sentry-sdk==0.10.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.3
werkzeug==0.15.4

# The following packages are considered to be unsafe in a requirements file:
# pip==19.1.1               # via pip-review
